+++ avpz_back/src/main/java/lpnu/vlpi/avpz/service/BaseService.java
@@ -2,5 +2,5 @@ package lpnu.vlpi.avpz.service;
 
 public interface BaseService {
     String getNewUid();
-    int getPagesCount(int size);
+    int getPagesCount(long size);
 }
