+++ avpz_back/src/main/java/lpnu/vlpi/avpz/service/impl/DefaultTaskService.java
@@ -49,7 +49,7 @@ public class DefaultTaskService implements TaskService {
     }
 
     @Override
-    public int getPagesCount(int size) {
+    public int getPagesCount(long size) {
         return 0;
     }
 }
