+++ avpz_back/src/main/java/lpnu/vlpi/avpz/service/impl/DefaultUserService.java
@@ -86,7 +86,7 @@ public class DefaultUserService implements UserService {
     }
 
     @Override
-    public int getPagesCount(int size) {
+    public int getPagesCount(long size) {
         return userRepository.getTotalPages(size);
     }
 }
